requirements.txt
aiogram>=3.20.0
aiohttp>=3.11.18
python-dotenv>=1.1.0